nanoclaw_service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
anthropic>=0.28.0
pyyaml>=6.0.1
redis>=5.0.4
aiosqlite>=0.20.0
pydantic>=2.7.0
python-dotenv>=1.0.1
